Mejoras en el proceso de AnexosSir

+++ ws/regweb3_api/src/test/java/es/caib/regweb3/ws/v3/test/RegWebRegistroEntradaTest.java
@@ -280,7 +280,7 @@ public class RegWebRegistroEntradaTest extends RegWebTestUtils {
         anexoWs.setValidezDocumento(CODIGO_SICRES_BY_TIPOVALIDEZDOCUMENTO.get(TIPOVALIDEZDOCUMENTO_COPIA));
         anexoWs.setTipoDocumental(getTestAnexoTipoDocumental());
         anexoWs.setTipoDocumento(CODIGO_SICRES_BY_TIPO_DOCUMENTO.get(TIPO_DOCUMENTO_FORMULARIO));
-        anexoWs.setOrigenCiudadanoAdmin(ANEXO_ORIGEN_CIUDADANO.intValue());
+        anexoWs.setOrigenCiudadanoAdmin(ANEXO_ORIGEN_CIUDADANO);
         anexoWs.setObservaciones("Observaciones de Marilen");
 
         anexoWs.setModoFirma(MODO_FIRMA_ANEXO_SINFIRMA);
@@ -313,7 +313,7 @@ public class RegWebRegistroEntradaTest extends RegWebTestUtils {
         anexoFirmaAtached.setValidezDocumento(CODIGO_SICRES_BY_TIPOVALIDEZDOCUMENTO.get(TIPOVALIDEZDOCUMENTO_COPIA_COMPULSADA));
         anexoFirmaAtached.setTipoDocumental(getTestAnexoTipoDocumental());
         anexoFirmaAtached.setTipoDocumento(CODIGO_SICRES_BY_TIPO_DOCUMENTO.get(TIPO_DOCUMENTO_DOC_ADJUNTO));
-        anexoFirmaAtached.setOrigenCiudadanoAdmin(ANEXO_ORIGEN_ADMINISTRACION.intValue());
+        anexoFirmaAtached.setOrigenCiudadanoAdmin(ANEXO_ORIGEN_ADMINISTRACION);
         anexoFirmaAtached.setObservaciones("Observaciones firma atached");
         anexoFirmaAtached.setModoFirma(MODO_FIRMA_ANEXO_ATTACHED);
         // Fichero Anexado
@@ -342,7 +342,7 @@ public class RegWebRegistroEntradaTest extends RegWebTestUtils {
         anexoFirmaDetached.setValidezDocumento(CODIGO_SICRES_BY_TIPOVALIDEZDOCUMENTO.get(TIPOVALIDEZDOCUMENTO_COPIA_ORIGINAL));
         anexoFirmaDetached.setTipoDocumental(getTestAnexoTipoDocumental());
         anexoFirmaDetached.setTipoDocumento(CODIGO_SICRES_BY_TIPO_DOCUMENTO.get(TIPO_DOCUMENTO_FICHERO_TECNICO));
-        anexoFirmaDetached.setOrigenCiudadanoAdmin(ANEXO_ORIGEN_CIUDADANO.intValue());
+        anexoFirmaDetached.setOrigenCiudadanoAdmin(ANEXO_ORIGEN_CIUDADANO);
         anexoFirmaDetached.setObservaciones("Observaciones firma detached");
         anexoFirmaDetached.setModoFirma(MODO_FIRMA_ANEXO_DETACHED);
         // Fichero Anexado
@@ -380,7 +380,7 @@ public class RegWebRegistroEntradaTest extends RegWebTestUtils {
         anexoFirmaDetachedCopia.setValidezDocumento(CODIGO_SICRES_BY_TIPOVALIDEZDOCUMENTO.get(TIPOVALIDEZDOCUMENTO_COPIA_COMPULSADA));
         anexoFirmaDetachedCopia.setTipoDocumental(getTestAnexoTipoDocumental());
         anexoFirmaDetachedCopia.setTipoDocumento(CODIGO_SICRES_BY_TIPO_DOCUMENTO.get(TIPO_DOCUMENTO_FORMULARIO));
-        anexoFirmaDetachedCopia.setOrigenCiudadanoAdmin(ANEXO_ORIGEN_CIUDADANO.intValue());
+        anexoFirmaDetachedCopia.setOrigenCiudadanoAdmin(ANEXO_ORIGEN_CIUDADANO);
         anexoFirmaDetachedCopia.setObservaciones("Observaciones firma detached copia");
         anexoFirmaDetachedCopia.setModoFirma(MODO_FIRMA_ANEXO_DETACHED);
         // Fichero Anexado
